requirements.txt
PyYAML~=6.0.2
pvlib~=0.13.1
pandas~=2.2.2
numpy~=1.26.4
numpy-financial
matplotlib~=3.9.2
soupsieve~=2.8
beautifulsoup4~=4.14.2
typing_extensions~=4.12.2
lxml~=5.3.0
h11~=0.16.0
defusedxml~=0.7.1
ipython~=9.6.0
pillow~=10.4.0
jupyter~=1.1.1
pip~=24.0
attrs~=25.3.0
wheel~=0.45.1
tornado~=6.5.2
Jinja2~=3.1.6
pyzmq~=27.1.0
colorama~=0.4.6
cffi~=2.0.0
setuptools~=74.1.2
comm~=0.2.3
fqdn~=1.5.1
h5py~=3.11.0
parso~=0.8.5
jedi~=0.19.2
lark~=1.3.0
rich~=13.8.0
pytz~=2024.1
sniffio~=1.3.1
arrow~=1.4.0
python-dateutil~=2.9.0.post0
babel~=2.17.0
click~=8.2.1
Pygments~=2.18.0
httpcore~=1.0.9
idna~=3.8
httpx~=0.28.1
json5~=0.12.1
scipy~=1.14.1
requests~=2.32.3
threadpoolctl~=3.5.0
packaging~=24.1
pooch~=1.8.2
platformdirs~=4.2.2
bleach~=6.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
MarkupSafe~=3.0.2
openpyxl~=3.1.5
SQLAlchemy~=2.0.40
psutil~=7.1.1
certifi~=2024.8.30
debugpy~=1.8.17
traitlets~=5.14.3
greenlet~=3.2.1
decorator~=5.2.1
ipython_pygments_lexers~=1.1.1
jupyter_client~=8.6.3
prompt_toolkit~=3.0.52
wcwidth~=0.2.14
Markdown~=3.8
urllib3~=2.2.2
six~=1.16.0
anyio~=4.11.0
nbformat~=5.10.4
jupyter_core~=5.9.1
nbclient~=0.10.2
fastjsonschema~=2.21.2
jsonschema~=4.25.1
jupyter_server~=2.17.0
jupyterlab~=4.4.9
jupyterlab_server~=2.27.3
notebook_shim~=0.2.4
notebook~=7.4.7
et_xmlfile~=2.0.0
asttokens~=3.0.0
contourpy~=1.3.0
executing~=2.2.1
fonttools~=4.53.1
ipykernel~=7.0.1
nbconvert~=7.16.6
pandocfilters~=1.5.1
mistune~=3.1.4
jupyterlab_pygments~=0.3.0
pure_eval~=0.2.3
pycparser~=2.23
pyparsing~=3.1.4
websocket-client~=1.9.0
terminado~=0.18.1
ipywidgets~=8.1.7
referencing~=0.37.0
webcolors~=24.11.1
jsonpointer~=3.0.0
isoduration~=20.11.0
cycler~=0.12.1
kiwisolver~=1.4.7
Send2Trash~=1.8.3
uritemplate~=3.0.1
google-api-python-client~=2.182.0
prometheus_client~=0.23.1
jupyter_server_terminals~=0.5.3
Werkzeug~=3.1.3